requirements.txt
djangorestframework-simplejwt == 5.5.1